rs_ai/requirements.txt
flask==3.0.0
flask-cors==5.0.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.5.0
mistralai==0.1.5
newsapi-python==0.2.7